Update ArimBans to reference refactored ArimAPI
