+++ ArimBans3/pom.xml
@@ -115,6 +115,10 @@
 	<dependencies>
 		<dependency>
 			<groupId>space.arim</groupId>
+			<artifactId>arimapi-framework</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>space.arim</groupId>
 			<artifactId>bans-resources</artifactId>
 			<classifier>resources</classifier>
 			<type>zip</type>
