+++ ArimBans3/src/main/java/space/arim/bans/internal/frontend/format/Formats.java
@@ -37,8 +37,8 @@ import space.arim.bans.api.CommandType.SubCategory;
 import space.arim.bans.api.exception.InternalStateException;
 import space.arim.bans.api.exception.InvalidSubjectException;
 
+import space.arim.api.framework.PlayerNotFoundException;
 import space.arim.api.util.StringsUtil;
-import space.arim.api.uuid.PlayerNotFoundException;
 
 public class Formats implements FormatsMaster {
