+++ pom.xml
@@ -102,6 +102,11 @@
 			</dependency>
 			<dependency>
 				<groupId>space.arim</groupId>
+				<artifactId>arimapi-framework</artifactId>
+				<version>${arimapi.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>space.arim</groupId>
 				<artifactId>universal-registry</artifactId>
 				<version>${universal.version}</version>
 			</dependency>
